Add for the view meal and delete meal handlers
Functionality add to ensure system is capable of displaying detailed
meal information to the user. The user can also delete the meal they
are viewing from the system. Future iterations should layer in logic to
only allow a user to delete if they are logged in and the initial
meal/recipe creator.